refactor: 연관관계 테스트 클래스 이름 변경
