+++ src/test/java/com/programmers/jpa/OrderItemAssociationTest.java
@@ -1,9 +1,13 @@
-package com.programmers.jpa.repository;
+package com.programmers.jpa;
 
 import com.programmers.jpa.domain.Item;
 import com.programmers.jpa.domain.Member;
 import com.programmers.jpa.domain.Order;
 import com.programmers.jpa.domain.OrderItem;
+import com.programmers.jpa.repository.ItemRepository;
+import com.programmers.jpa.repository.MemberRepository;
+import com.programmers.jpa.repository.OrderItemRepository;
+import com.programmers.jpa.repository.OrderRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import org.junit.jupiter.api.BeforeEach;
@@ -15,7 +19,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
-class OrderItemRepositoryTest {
+class OrderItemAssociationTest {
     @Autowired
     MemberRepository memberRepository;
     @Autowired
